FIXX

+++ backend/app/src/main/java/exchangerate/ExchangeRateResponse.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 public class ExchangeRateResponse {
     private String baseCurrency;
-    private Map<String, BigDecimal> conversionRates;
+    private Map<String, Double> conversionRates;
 
     public ExchangeRateResponse(String baseCurrency, Map<String, Double> conversionRates) {
         this.baseCurrency = baseCurrency;
